Move analyze-storage over to blockstore subcommand
